+++ publisher-sdk/src/androidTest/java/com/criteo/publisher/CriteoInterstitialActivityTest.java
@@ -26,6 +26,7 @@ import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.clearInvocations;
 import static org.mockito.Mockito.inOrder;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
@@ -176,6 +177,7 @@ public class CriteoInterstitialActivityTest {
   private CriteoInterstitialActivity givenOpenedInterstitialActivity(@NonNull String html) throws Exception {
     Activity activity = lookup.lookForResumedActivity(() -> {
       InterstitialListenerNotifier listenerNotifier = new InterstitialListenerNotifier(
+          mock(CriteoInterstitial.class),
           listener,
           runOnUiThreadExecutor
       );