+++ publisher-sdk/src/main/java/com/criteo/publisher/CriteoInterstitial.java
@@ -126,6 +126,7 @@ public class CriteoInterstitial {
       Criteo criteo = getCriteo();
 
       InterstitialListenerNotifier listenerNotifier = new InterstitialListenerNotifier(
+          this,
           criteoInterstitialAdListener,
           getRunOnUiThreadExecutor()
       );