+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next version
+# 3.10.0
 
 - Features
   - MoPub Header-Bidding: Handle subclasses of `MoPubView` and `MoPubInterstitial`